requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
pydantic
pydantic-settings
SQLAlchemy
passlib
python-dotenv
alembic
psycopg2
starlette
argon2-cffi
PyJWT
httpx
pytest
allure-pytest
pytest
httpx
google-auth
requests